[Core] Allow a solution to be monitored by the file watcher
Added a simple way to register a solution so all files in its
directories are monitored. Only file change events are passed to the
FileService which are then raised as events by the FileService.
Added some tests to verify that modify a project inside a solution
or a file inside the project generates FileService FileChanged
events. The tests that use Project.SaveAsync pass without the
new file watcher since the FileService is called directly to
notify file changes when SolutionItem.DoSave is called and
MSBuildProject.SaveAsync is called, which uses TextFile.Write which
generates the event. The direct calls to notify the FileService
have not yet been removed.