[Ide] Use file watcher to detect changes instead of checking all files
Removed the code that checked all the files belonging to open solution
when the IDE lost focus, or a build was started, then checked for
modified or removed files on regaining focus or the build completes.
This is now handled by the file watcher for the solution. Note that
in the case of the IDE losing focus the file events are frozen until
the IDE gets focus again.