[Core] Watch files outside the solution directory
Handle projects and files that exist outside the solution directory
when watching files. File watchers are now created for all directories
that are needed instead of just for the solution's directory.
Currently the directory information is cached in the WorkspaceItem
and not updated when a new project or file is added or removed.