[Core] Fix file removed event fired when file not deleted
The native file watcher sometimes generates Changed, Created and
Deleted events in that order from a single native file event. This
can occur even when the file is not deleted. Sometimes you can
trigger this behaviour using File.WriteAllText when the file does
not exist. Now a check is made to ensure the file does not exist
before the file watcher deleted event handler raises the FileRemoved
event with the FileService.